+++ app/code/Ecomail/Ecomail/view/frontend/requirejs-config.js
@@ -1,0 +1,7 @@
+var config = {
+    map: {
+        '*': {
+            'ecomail': 'Ecomail_Ecomail/js/tracking'
+        }
+    }
+};