+++ app/code/Ecomail/Ecomail/CHANGELOG.md
@@ -5,3 +5,7 @@
 
 ## v2.1.0 - 2 Jun, 2023
 * added Magento 2.4.6 compatibility
+
+## v2.2.0 - 2 Jun, 2023
+* added webhook for subscriber status synchronization
+* fixed issue with duplicated subscribers